gRpc_project/requirements.txt
sqlalchemy>=2.0
asyncpg
grpcio
grpcio-tools
pytest
pytest-asyncio
python-dotenv
